Programmatically determining root view controller
Setting up the app to have the root view controller be set up in the app
delegate code rather than having the view controller be defined in a
storyboard.

Also a bunch of project management in the xcodeproj and other files.
Normally I'd go back and patch individual line changes in on the commits
they are relevant in, but this is just a fun project and I have limited
time on earth to figure out how to edit history in git again.