Showing checked answers in game view
Needed to move some stuff around between GameView and GameViewController
for the responsibilties of these classes to make more sense. Moved the
text field out of the GameView and into the VC, and set up the VC so
that it check and submit guesses to be displayed in the game view.